chore: organize source folders
