+++ README.md
@@ -14,18 +14,7 @@ npm run dev
 ```
 
 ### Offline Caching
-
-<<<<<<< HEAD
-The client caches recent master data (courses, educations, exams, batches and payment modes)
-inside **IndexedDB** using [Dexie](https://dexie.org). Sensitive fields are encrypted via
-`crypto-js` before being stored. All cached data is purged on user logout.
-=======
-The client caches leads, students, attendance, admissions, courses, exams,
-batches and payment modes in **IndexedDB** via [Dexie](https://dexie.org).
-Highly sensitive fields are encrypted with `crypto-js` before storage.
-Set `VITE_DB_SECRET_KEY` in your environment files to configure encryption.
-All cached data is automatically purged on logout.
->>>>>>> origin/5toixc-codex/implement-indexeddb-with-dexie-and-encryption
+The client caches leads, students, attendance, admissions, courses, exams, batches and payment modes in **IndexedDB** via [Dexie](https://dexie.org). Sensitive fields are encrypted with `crypto-js` before storage and purged on logout. Set `VITE_DB_SECRET_KEY` in your environment files to configure encryption.
 
 ## API Endpoints
